Add file for router, different file naming, removing useless string
