+++ client/src/pages/GenreDetail.css
@@ -12,7 +12,7 @@
   opacity: 0.9;
 }
 
-.genre-header-title {
+.genre-header h1 {
   font-size: 5rem;
   margin-left: 2rem;
 }
